LayoutTests/inspector/editor/text-editor-highlight-regexp-expected.txt

Test editor highlight regexp API.


Running: testHighlightApi

Before highlight

<div class="inner-container" tabindex="0">
<div class="webkit-line-content">foo();</div>
<div class="webkit-line-content">foo_1; foor; foo_;</div>
<div class="webkit-line-content">foo_</div>
<div class="webkit-line-content">global_foo</div>
<div class="webkit-line-content">global_foo2</div>
<div class="webkit-line-content">some_other_foo_X</div></div>

Add highlight for regexp /foo_./

<div class="inner-container" tabindex="0">
<div class="webkit-line-content">foo();</div>
<div class="webkit-line-content">foo_1; foor; foo_;<span class="some-css-class text-editor-overlay-highlight" style="margin-left: -127px; width: 37px;"></span><span class="some-css-class text-editor-overlay-highlight" style="margin-left: -36px; width: 37px;"></span></div>
<div class="webkit-line-content">foo_</div>
<div class="webkit-line-content">global_foo</div>
<div class="webkit-line-content">global_foo2</div>
<div class="webkit-line-content">some_other_foo_X<span class="some-css-class text-editor-overlay-highlight" style="margin-left: -36px; width: 37px;"></span></div></div>

Remove highlight for regexp /foo_./

<div class="inner-container" tabindex="0">
<div class="webkit-line-content">foo();</div>
<div class="webkit-line-content">foo_1; foor; foo_;</div>
<div class="webkit-line-content">foo_</div>
<div class="webkit-line-content">global_foo</div>
<div class="webkit-line-content">global_foo2</div>
<div class="webkit-line-content">some_other_foo_X</div></div>
